@@ -13,48 +13,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.1.0",
-    "@angular/cdk": "^10.2.0",
-    "@angular/common": "^10.1.0",
-    "@angular/compiler": "^10.1.0",
-    "@angular/core": "^10.1.0",
-    "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "^10.1.0",
-    "@angular/material": "^10.2.0",
-    "@angular/platform-browser": "^10.1.0",
-    "@angular/platform-browser-dynamic": "^10.1.0",
-    "@angular/router": "^10.1.0",
-    "@swimlane/ngx-charts": "^16.0.0",
-    "ajv": "^6.12.4",
-    "angular-oauth2-oidc": "^10.0.3",
+    "@angular/animations": "^12.2.7",
+    "@angular/cdk": "^12.2.7",
+    "@angular/common": "^12.2.7",
+    "@angular/compiler": "^12.2.7",
+    "@angular/core": "^12.2.7",
+    "@angular/flex-layout": "^12.0.0-beta.35",
+    "@angular/forms": "^12.2.7",
+    "@angular/material": "^12.2.7",
+    "@angular/platform-browser": "^12.2.7",
+    "@angular/platform-browser-dynamic": "^12.2.7",
+    "@angular/router": "^12.2.7",
+    "@swimlane/ngx-charts": "^19.0.1",
+    "ajv": "^6.12.6",
+    "angular-oauth2-oidc": "^12.1.0",
     "csv-rxjs-kit": "^1.2.0",
     "ez-jwt": "^1.1.2",
     "jsvrx-ajv": "^1.0.2",
-    "rxjs": "^6.6.3",
-    "rxjs-compat": "^6.6.3",
-    "zone.js": "^0.10.3"
+    "rxjs": "^7.3.0",
+    "zone.js": "^0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1001.0",
-    "@angular-devkit/schematics": "^10.1.0",
+    "@angular-devkit/build-angular": "^12.2.7",
+    "@angular-devkit/schematics": "^12.2.7",
     "@angular/cli": "^12.2.7",
-    "@angular/compiler-cli": "^10.1.0",
-    "@angular/language-service": "^10.1.0",
-    "@types/djv": "^2.1.0",
-    "@types/json-schema": "^7.0.6",
-    "@types/node": "^14.6.4",
-    "@typescript-eslint/eslint-plugin": "^4.0.1",
-    "@typescript-eslint/parser": "^4.0.1",
+    "@angular/compiler-cli": "^12.2.7",
+    "@angular/language-service": "^12.2.7",
+    "@types/djv": "^2.1.1",
+    "@types/json-schema": "^7.0.9",
+    "@types/node": "^16.10.1",
+    "@typescript-eslint/eslint-plugin": "^4.32.0",
+    "@typescript-eslint/parser": "^4.32.0",
     "djv": "^2.1.4",
-    "eslint": "^7.8.1",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-deprecation": "^1.1.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-deprecation": "^1.2.1",
     "eslint-plugin-no-null": "^1.0.2",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.20.6",
-    "prettier": "^2.1.1",
-    "source-map-explorer": "^2.5.0",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.2"
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.26.0",
+    "prettier": "^2.4.1",
+    "source-map-explorer": "^2.5.2",
+    "ts-node": "^10.2.1",
+    "typescript": "~4.3.2"
   }
 }
